# backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "QrBites API",
    "version": "1.0.0",
    "description": "\nQrBites Restaurant Menu Management System API Documentation.\n\n## Features\n- 🍽️ Digital menu management\n- 📱 QR code generation for menus\n- 🔐 JWT-based authentication\n- 👥 Multi-role authorization (Admin, Restaurant Owner, Customer)\n- 🖼️ Image upload with Cloudinary integration\n- 🔍 Advanced menu search and filtering\n\n## Authentication\nMost endpoints require JWT authentication. To authenticate:\n1. Register or login via /api/auth/register or /api/auth/login\n2. Copy the token from the response\n3. Click the \"Authorize\" button (🔒 icon) above\n4. Enter: `Bearer <your-token>`\n5. Click \"Authorize\" and \"Close\"\n\nYour token will be automatically included in subsequent requests.\n      ",
    "contact": {
      "name": "QrBites Support",
      "email": "support@qrbites.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://qrbites-api.vercel.app",
      "description": "Production (Vercel)"
    },
    {
      "url": "http://localhost:5000",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer {token}"
      }
    },
    "schemas": {
...
      "MenuItem": {
        "type": "object",
        "required": ["name", "price", "menuId"],
        "properties": {
          "_id": {
            "type": "string",
            "format": "objectId",
            "example": "507f1f77bcf86cd799439011"
          },
          "name": {
            "type": "string",
            "example": "Margherita Pizza"
          },
          "description": {
            "type": "string",
            "example": "Classic pizza with fresh mozzarella, tomato sauce, and basil"
          },
          "price": {
            "type": "number",
            "format": "float",
            "example": 12.99
          },
          "currency": {
            "type": "string",
            "example": "USD",
            "default": "USD"
          },
          "category": {
            "type": "string",
            "example": "Main Course"
          },
          "menuId": {
            "type": "string",
            "format": "objectId",
            "example": "507f1f77bcf86cd799439011"
          },
          "image": {
            "type": "string",
            "format": "uri",
            "example": "https://res.cloudinary.com/xxx/image/upload/v123/pizza.jpg"
          },
          "isAvailable": {
            "type": "boolean",
            "example": true
          },
          "allergens": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["gluten", "dairy"]
          },
          "dietary": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["vegetarian"]
          },
          "spicyLevel": {
            "type": "number",
            "minimum": 0,
            "maximum": 5,
            "example": 2
          },
          "preparationTime": {
            "type": "number",
            "description": "Preparation time in minutes",
            "example": 15
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
...
    "parameters": {
...
      "PageParam": {
        "name": "page",
        "in": "query",
        "description": "Page number for pagination (starts at 1)",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1,
          "example": 1
        }
      },
      "LimitParam": {
        "name": "limit",
        "in": "query",
        "description": "Number of items per page",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 10,
          "example": 10
        }
      },
      "SortParam": {
        "name": "sort",
        "in": "query",
        "description": "Sort field and order (prefix with - for descending)",
        "schema": {
          "type": "string",
          "example": "-createdAt"
        }
      },
      "SearchParam": {
        "name": "search",
        "in": "query",
        "description": "Search query string",
        "schema": {
          "type": "string",
          "example": "pizza"
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/menu-items": {
      "get": {
        "summary": "Get all menu items",
        "description": "Retrieve a list of menu items with optional filtering. Restaurant owners see only their menu items.",
        "tags": ["Menu Items"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          },
          {
            "$ref": "#/components/parameters/SortParam"
          },
          {
            "$ref": "#/components/parameters/SearchParam"
          },
          {
            "name": "menuId",
            "in": "query",
            "description": "Filter by menu ID",
            "schema": {
              "type": "string",
              "format": "objectId"
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Filter by category",
            "schema": {
              "type": "string",
              "example": "Appetizers"
            }
          },
          {
            "name": "isAvailable",
            "in": "query",
            "description": "Filter by availability status",
            "schema": {
              "type": "boolean",
              "example": true
            }
          },
          {
            "name": "minPrice",
            "in": "query",
            "description": "Minimum price filter",
            "schema": {
              "type": "number",
              "example": 5
            }
          },
          {
            "name": "maxPrice",
            "in": "query",
            "description": "Maximum price filter",
            "schema": {
              "type": "number",
              "example": 20
            }
          },
          {
            "name": "dietary",
            "in": "query",
            "description": "Filter by dietary preference (e.g., vegetarian, vegan)",
            "schema": {
              "type": "string",
              "example": "vegetarian"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved menu items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "menuItems": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/MenuItem"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/Pagination"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "summary": "Create a new menu item",
        "description": "Create a new menu item with optional image upload. Restaurant owners can only create items for their own menus.",
        "tags": ["Menu Items"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["name", "price", "menuId"],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Margherita Pizza",
                    "description": "Item name"
                  },
                  "description": {
                    "type": "string",
                    "example": "Classic pizza with fresh mozzarella, tomato sauce, and basil"
                  },
                  "price": {
                    "type": "number",
                    "format": "float",
                    "example": 12.99,
                    "description": "Item price"
                  },
                  "currency": {
                    "type": "string",
                    "example": "USD",
                    "default": "USD"
                  },
                  "category": {
                    "type": "string",
                    "example": "Main Course"
                  },
                  "menuId": {
                    "type": "string",
                    "format": "objectId",
                    "example": "507f1f77bcf86cd799439011",
                    "description": "Menu ID (must be owned by user)"
                  },
                  "isAvailable": {
                    "type": "boolean",
                    "example": true,
                    "default": true
                  },
                  "allergens": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": ["gluten", "dairy"],
                    "description": "List of allergens (JSON array as string)"
                  },
                  "dietary": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": ["vegetarian"],
                    "description": "Dietary tags (JSON array as string)"
                  },
                  "spicyLevel": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 5,
                    "example": 2
                  },
                  "preparationTime": {
                    "type": "number",
                    "example": 15,
                    "description": "Preparation time in minutes"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Menu item image"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Menu item created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/MenuItem"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/menu-items/{id}": {
      "get": {
        "summary": "Get menu item by ID",
        "description": "Retrieve a specific menu item by its ID. Users can only access items from their own menus.",
        "tags": ["Menu Items"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Menu item retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/MenuItem"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "put": {
        "summary": "Update a menu item",
        "description": "Update an existing menu item. Only the owner can update their menu items.",
        "tags": ["Menu Items"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Updated Margherita Pizza"
                  },
                  "description": {
                    "type": "string",
                    "example": "Now with extra basil!"
                  },
                  "price": {
                    "type": "number",
                    "format": "float",
                    "example": 13.99
                  },
                  "currency": {
                    "type": "string",
                    "example": "USD"
                  },
                  "category": {
                    "type": "string",
                    "example": "Main Course"
                  },
                  "isAvailable": {
                    "type": "boolean",
                    "example": true
                  },
                  "allergens": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": ["gluten", "dairy"]
                  },
                  "dietary": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": ["vegetarian"]
                  },
                  "spicyLevel": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 5,
                    "example": 2
                  },
                  "preparationTime": {
                    "type": "number",
                    "example": 15
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "New menu item image (replaces existing)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Menu item updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/MenuItem"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "delete": {
        "summary": "Delete a menu item",
        "description": "Delete a menu item permanently. Only the owner can delete their menu items.",
        "tags": ["Menu Items"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdParam"
          }
        ],
        "responses": {
          "204": {
            "description": "Menu item deleted successfully (no content)"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/menu-items/{id}/image": {
      "post": {
        "summary": "Upload menu item image",
        "description": "Upload or replace the image for a menu item. The image will be uploaded to Cloudinary.",
        "tags": ["Menu Items"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["image"],
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Menu item image file (JPG, PNG, WEBP)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Image uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "imageUrl": {
                          "type": "string",
                          "format": "uri",
                          "example": "https://res.cloudinary.com/xxx/image/upload/v123/menuitem.jpg"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/api/menus/{id}/image": {
      "post": {
        "summary": "Upload menu image",
        "description": "Upload a single image for a menu. The image will be uploaded to Cloudinary.",
        "tags": ["Menus"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["image"],
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Menu image file"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Image uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "imageUrl": {
                          "type": "string",
                          "format": "uri",
                          "example": "https://res.cloudinary.com/xxx/image/upload/v123/menu.jpg"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  }
}